refactoring team list 1
