+++ lib/screens/home_screen.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:my_softball_team/widgets/season_schedule.dart';
-import 'package:my_softball_team/widgets/team_list.dart';
-import 'package:my_softball_team/screens/Stats Leaderboard/leaderboard_home.dart';
+import 'package:my_softball_team/screens/TeamList/team_list.dart';
+import 'package:my_softball_team/screens/StatsLeaderboard/leaderboard_home.dart';
 import 'package:my_softball_team/globals.dart' as globals;
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:groovin_material_icons/groovin_material_icons.dart';
@@ -282,7 +282,7 @@ class _HomeScreenState extends State<HomeScreen> {
         children: <Widget>[
           SeasonSchedule(),
           TeamList(),
-          StatsTable()
+          LeaderboardHome()
         ],
         controller: _pageController,
         onPageChanged: _onPageChanged,
